refactor: split region map political render into label and border parts (#3878)
Decompose the 641-line political render part into province/sea-zone label
plates and political border overlays so each region_map submodule stays under
the Phase 3 flame 600-line target without changing map behavior.

Refs #3878
